+++ src/routes/resources/animtab/index.tsx
@@ -1,20 +1,20 @@
-import { component$, useSignal, useStore, useTask$, useVisibleTask$ } from '@builder.io/qwik';
+import { component$, useStore, useTask$, useVisibleTask$ } from '@builder.io/qwik';
 import { routeLoader$, type DocumentHead } from '@builder.io/qwik-city';
 
-import { defaults, loadPreset, types, v3formats, presets as presetlist } from '~/components/util/PresetUtils';
-import { AnimationOutput, convertToHex, convertToRGB, getAnimFrames, hexToHSL } from '~/components/util/RGBUtils';
+import { defaults, types, v3formats } from '~/components/util/PresetUtils';
+import { AnimationOutput, getAnimFrames, hexToHSL } from '~/components/util/RGBUtils';
 
 import { Dropdown, Toggle, NumberInput } from '@luminescent/ui-qwik';
 import { inlineTranslate, useSpeak } from 'qwik-speak';
-import { getCookies, setCookies, sortColors } from '~/components/util/SharedUtils';
+import { getCookies, setCookies } from '~/components/util/SharedUtils';
 import { isBrowser } from '@builder.io/qwik/build';
 import { rgbDefaults } from '../rgb';
-import { Gradient } from '~/components/util/HexUtils';
 import Input from '~/components/rgb/Input';
 import ColorMap from '~/components/rgb/ColorMap';
 import ColorList from '~/components/rgb/ColorList';
-import { ChevronDown, Clipboard, Palette } from 'lucide-icons-qwik';
+import { ChevronDown, Clipboard, Palette, Save } from 'lucide-icons-qwik';
 import Output from '~/components/rgb/Output';
+import Presets from '~/components/rgb/Presets';
 
 export const animTABDefaults = {
   name: defaults.name,
@@ -80,8 +80,6 @@ export default component$(() => {
     frames: [],
     frame: 0,
   }, { deep: true });
-
-  const modalRef = useSignal<HTMLDialogElement>();
 
   // eslint-disable-next-line qwik/no-use-visible-task
   useVisibleTask$(async () => {
@@ -335,230 +333,23 @@ export default component$(() => {
                 </div>
               </>
             }
-            <div class="flex flex-col gap-2 mt-4">
-              <h1 class="hidden sm:flex md:text-lg xl:text-xl font-semibold text-gray-50 gap-3 items-center justify-center">
+            <button class="lum-btn lum-bg-gray-800/30 rounded-md lum-pad-md" onClick$={() => {
+              if (tmpstore.sectionsOpened.indexOf('presets') == -1) tmpstore.sectionsOpened.push('presets');
+              else tmpstore.sectionsOpened.splice(tmpstore.sectionsOpened.indexOf('presets'), 1);
+            }}>
+              <h1 class="flex flex-1 md:text-lg xl:text-xl font-semibold text-gray-50 gap-3 items-center">
+                <Save size={26} />
                 {t('color.presets@@Presets')}
               </h1>
-              <div class="grid grid-cols-2 gap-2">
-                <div class="flex flex-col gap-2">
-                  <Dropdown id="saved-presets" class={{ 'w-full': true }} onChange$={
-                    (event, el) => {
-                      let json: Partial<typeof defaults> = {};
-                      try {
-                        const preset = loadPreset(el.value);
-                        navigator.clipboard.writeText(JSON.stringify(preset));
-                        json = {
-                          ...preset,
-                        };
-                      } catch (err) {
-                        const alert = {
-                          class: 'text-red-500',
-                          text: 'color.invalidPreset@@INVALID PRESET! Please report this to the <a class="text-blue-400 hover:underline" href="https://discord.gg/9vUZ9MREVz">Developers</a> with the preset you tried to import.',
-                        };
-                        const errtext = {
-                          class: 'text-red-300',
-                          text: `${err}`,
-                        };
-                        tmpstore.alerts.push(alert, errtext);
-                        return setTimeout(() => {
-                          tmpstore.alerts.splice(tmpstore.alerts.indexOf(alert), 1);
-                          tmpstore.alerts.splice(tmpstore.alerts.indexOf(errtext), 1);
-                        }, 5000);
-                      }
-                      (Object.keys(store) as Array<keyof typeof store>).forEach(key => {
-                        if (store[key] === undefined) return;
-                        (store as any)[key] = json[key] ?? defaults[key];
-                      });
-                      (Object.keys(animtabstore) as Array<keyof typeof animtabstore>).forEach(key => {
-                        if (animtabstore[key] === undefined) return;
-                        (animtabstore as any)[key] = json[key] ?? defaults[key];
-                      });
-                      const alert = {
-                        class: 'text-green-500',
-                        text: 'color.importedPreset@@Successfully imported preset!',
-                      };
-                      tmpstore.alerts.push(alert);
-                      setTimeout(() => {
-                        tmpstore.alerts.splice(tmpstore.alerts.indexOf(alert), 1);
-                      }, 2000);
-                    }
-                  } values={
-                    presetstore.savedPresets.map((preset) => ({
-                      name: <span class={{
-                        'break-all font-mc tracking-tight': true,
-                      }}>
-                        {(() => {
-                          const colors = sortColors(preset.colors ?? presetlist[0].colors).map((color) => ({ rgb: convertToRGB(color.hex), pos: color.pos }));
-                          if (colors.length < 2) return preset.name;
-
-                          const gradient = new Gradient(colors, Math.ceil((preset.name ?? 'Untitled').length));
-
-                          let hex = '';
-                          const segments = [...(preset.name ?? 'Untitled').matchAll(new RegExp('.{1,1}', 'g'))];
-                          return segments.map((segment, i) => {
-                            hex = convertToHex(gradient.next());
-                            return (
-                              <span key={`segment-${i}`} style={`color: #${hex};`}>
-                                {segment[0].replace(/ /g, '\u00A0')}
-                              </span>
-                            );
-                          });
-                        })()}
-                      </span>,
-                      value: JSON.stringify(preset),
-                    }))
-                  } display={<span class="flex gap-3 flex-1">
-                    Load saved preset
-                  </span>}>
-                    {t('color.savedPresets@@Saved Presets')}
-                  </Dropdown>
-                  <div class="grid grid-cols-2 gap-2">
-                    <a class="lum-btn" href="/resources/rgb/presets">
-                      Browse
-                    </a>
-                    <button class="lum-btn" id="save" onClick$={() => {
-                      modalRef.value?.showModal();
-                    }}>
-                      {t('color.save@@Save')}
-                    </button>
-                    <dialog ref={modalRef} class="lum-bg-gray-800/20 lum-pad-equal-2xl shadow-lg backdrop-blur-xl rounded-lg relative max-w-lg w-full transform transition-transform duration-300 ease-out">
-                      <div class="flex flex-col gap-3">
-                        <h3 class="text-gray-50 text-xl font-semibold mb-4">
-                          {t('color.savePreset@@Save Preset')}
-                        </h3>
-
-                        <input class="lum-input" id="presetname" placeholder={t('color.presetName@@Preset Name')} />
-
-                        <div class="flex gap-2 justify-end">
-                          <button class="lum-btn" onClick$={() => {
-                            modalRef.value?.close();
-                          }}>
-                            {t('color.cancel@@Cancel')}
-                          </button>
-                          <button class="lum-btn lum-bg-green-900 hover:lum-bg-green-800" id="save" onClick$={() => {
-                            const presetnameinput = document.getElementById('presetname') as HTMLInputElement;
-                            const preset: Partial<typeof defaults> = {
-                              ...store,
-                              ...animtabstore,
-                              name: presetnameinput.value ?? 'Untitled',
-                            };
-                            (Object.keys(preset) as Array<keyof typeof defaults>).forEach(key => {
-                              if (key != 'version' && JSON.stringify(preset[key]) === JSON.stringify(defaults[key as keyof typeof defaults])) delete preset[key];
-                            });
-                            if (presetstore.savedPresets.find(p => JSON.stringify(p) === JSON.stringify(preset))) return;
-                            presetstore.savedPresets.push(preset);
-                            setCookies('presets', presetstore);
-                            modalRef.value?.close();
-                            const alert = {
-                              class: 'text-green-500',
-                              text: 'color.savedPreset@@Successfully saved preset!',
-                            };
-                            tmpstore.alerts.push(alert);
-                            setTimeout(() => {
-                              tmpstore.alerts.splice(tmpstore.alerts.indexOf(alert), 1);
-                            }, 2000);
-                          }}>
-                            {t('color.save@@Save')}
-                          </button>
-                        </div>
-                      </div>
-                    </dialog>
-                  </div>
-                </div>
-                <div class="flex flex-col gap-2">
-                  <div class="flex flex-col gap-1">
-                    <label for="import">
-                      {t('color.import@@Import')}
-                    </label>
-                    <input class="lum-input" id="import" name="import" placeholder={t('color.import@@Import (Paste here)')} onInput$={async (e, el) => {
-                      let json: Partial<typeof defaults> = {};
-                      try {
-                        const preset = loadPreset(el.value);
-                        el.value = JSON.stringify(preset);
-                        navigator.clipboard.writeText(JSON.stringify(preset));
-                        json = {
-                          ...preset,
-                        };
-                      } catch (err) {
-                        const alert = {
-                          class: 'text-red-500',
-                          text: 'color.invalidPreset@@INVALID PRESET! Please report this to the <a class="text-blue-400 hover:underline" href="https://discord.gg/9vUZ9MREVz">Developers</a> with the preset you tried to import.',
-                        };
-                        const errtext = {
-                          class: 'text-red-300',
-                          text: `${err}`,
-                        };
-                        tmpstore.alerts.push(alert, errtext);
-                        return setTimeout(() => {
-                          tmpstore.alerts.splice(tmpstore.alerts.indexOf(alert), 1);
-                          tmpstore.alerts.splice(tmpstore.alerts.indexOf(errtext), 1);
-                        }, 5000);
-                      }
-                      (Object.keys(store) as Array<keyof typeof store>).forEach(key => {
-                        if (store[key] === undefined) return;
-                        (store as any)[key] = json[key] ?? defaults[key];
-                      });
-                      const alert = {
-                        class: 'text-green-500',
-                        text: 'color.importedPreset@@Successfully imported preset!',
-                      };
-                      tmpstore.alerts.push(alert);
-                      setTimeout(() => {
-                        tmpstore.alerts.splice(tmpstore.alerts.indexOf(alert), 1);
-                      }, 2000);
-                    }}/>
-                  </div>
-                  <div class="grid grid-cols-2 gap-2">
-                    <button class="lum-btn lum-pad-sm" id="export" onClick$={() => {
-                      const preset: Partial<typeof defaults> = { ...store, ...animtabstore };
-                      (Object.keys(preset) as Array<keyof typeof defaults>).forEach(key => {
-                        if (key != 'version' && JSON.stringify(preset[key]) === JSON.stringify(defaults[key as keyof typeof defaults])) delete preset[key];
-                      });
-                      navigator.clipboard.writeText(JSON.stringify(preset));
-                      const alert = {
-                        class: 'text-green-500',
-                        text: 'color.exportedPreset@@Successfully exported preset to clipboard!',
-                      };
-                      tmpstore.alerts.push(alert);
-                      setTimeout(() => {
-                        tmpstore.alerts.splice(tmpstore.alerts.indexOf(alert), 1);
-                      }, 2000);
-                    }}>
-                      {t('color.export@@Export')}
-                    </button>
-                    <button class="lum-btn lum-pad-sm" id="createurl" onClick$={() => {
-                      const base_url = `${window.location.protocol}//${window.location.host}${window.location.pathname}`;
-                      const url = new URL(base_url);
-                      const params: Partial<typeof defaults> = { ...store, ...animtabstore };
-                      (Object.entries(params) as Array<[keyof typeof defaults, any]>).forEach(([key, value]) => {
-                        if (key == 'format' || key == 'colors') {
-                          value = JSON.stringify(value);
-                          if (value === JSON.stringify(defaults[key as keyof typeof defaults])) return;
-                        }
-                        if (value === defaults[key as keyof typeof defaults]) return;
-                        url.searchParams.set(key, String(value));
-                      });
-                      window.history.pushState({}, '', url.href);
-                      const alert = {
-                        class: 'text-green-500',
-                        text: 'color.exportedPresetUrl@@Successfully exported preset to url!',
-                      };
-                      tmpstore.alerts.push(alert);
-                      setTimeout(() => {
-                        tmpstore.alerts.splice(tmpstore.alerts.indexOf(alert), 1);
-                      }, 2000);
-                    }}>
-                      {t('color.url@@Get URL')}
-                    </button>
-                  </div>
-                </div>
+              <div class={{
+                'transition-transform duration-200': true,
+                'rotate-180': tmpstore.sectionsOpened.indexOf('presets') != -1,
+              }}>
+                <ChevronDown size={20} />
               </div>
-              <div class="grid grid-cols-4 gap-2">
-              </div>
-              {tmpstore.alerts.map((alert, i) => (
-                <p key={`preset-alert${i}`} class={alert.class} dangerouslySetInnerHTML={t(alert.text)} />
-              ))}
-            </div>
+            </button>
+            <Presets store={store} presetstore={presetstore} tmpstore={tmpstore}
+              hidden={tmpstore.sectionsOpened.indexOf('presets') == -1}/>
           </div>
 
           <div class="flex flex-col gap-2" id="formatting">
